backend/requirements.txt
# FastAPI Backend Requirements
# VoiceAI Platform

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.6
flower==2.0.1

# OpenAI
openai==1.12.0
websockets==12.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Excel Processing
openpyxl==3.1.2
pandas==2.2.0

# File Storage
boto3==1.34.34  # MinIO S3 compatible
python-magic==0.4.27

# Audio Processing
pydub==0.25.1
numpy>=1.24.0

# Asterisk ARI
panoramisk==1.4

# Utilities
python-dotenv==1.0.1
orjson==3.9.12
phonenumbers==8.13.27
humanize==4.9.0

# Security
cryptography==42.0.2

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
ruff==0.1.14
